# pkg/docs/api/grpc/makosh-be_api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "api/grpc/makosh-be_api.proto",
    "version": "version not set"
  },
  "tags": [
    {
      "name": "makoshBeAPI"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/endpoints": {
      "post": {
        "operationId": "makoshBeAPI_UpsertEndpoints",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/makosh_be_apiUpsertEndpointsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/makosh_be_apiUpsertEndpointsRequest"
            }
          }
        ],
        "tags": [
          "makoshBeAPI"
        ]
      }
    },
    "/v1/endpoints/{serviceName}": {
      "get": {
        "operationId": "makoshBeAPI_ListEndpoints",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/makosh_be_apiListEndpointsResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "serviceName",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "makoshBeAPI"
        ]
      }
    },
    "/v1/version": {
      "get": {
        "operationId": "makoshBeAPI_Version",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/makosh_be_apiVersionResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "makoshBeAPI"
        ]
      }
    }
  },
  "definitions": {
    "makosh_be_apiEndpoint": {
      "type": "object",
      "properties": {
        "serviceName": {
          "type": "string"
        },
        "addrs": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "makosh_be_apiListEndpointsResponse": {
      "type": "object",
      "properties": {
        "urls": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "makosh_be_apiUpsertEndpointsRequest": {
      "type": "object",
      "properties": {
        "endpoints": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/makosh_be_apiEndpoint"
          }
        }
      }
    },
    "makosh_be_apiUpsertEndpointsResponse": {
      "type": "object"
    },
    "makosh_be_apiVersionResponse": {
      "type": "object",
      "properties": {
        "version": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    }
  }
}